requirements.txt
# Flask Web Framework
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==3.0.3

# Database
SQLAlchemy==2.0.35

# Web Server
gunicorn==23.0.0

# Scheduling
APScheduler==3.10.4

# Image Processing (for future enhancements)
Pillow==10.4.0

# Development and Utilities
python-dotenv==1.0.1